Add option to lign maps and bindings
This commit adds the option to allow maps and bindings.
It does so by adding two separate configurable options:

 - align-maps?

 - align-bindings?

Bindings specifically are further configurable by a cljfmt option:

 - align-bindings-args

This is a map of the form `{form #{1 2}}` that maps the form's expected
binding positions. An example is `{let #{0}}` implying that a `let`
symbol is immediately followed by a binding. The binding is expected
to be the first argument.