Add option to align maps and other forms
This commit adds the option to align maps and other forms.
It does so by adding two separate configurable options:

 - align-maps?

 - align-forms?

Forms specifically are further configurable by a cljfmt option:

 - aligns

This is a map of the form `{form #{1 2}}` that maps the form's expected
alignable positions. An example is `{let #{0}}` implying that a `let`
symbol is immediately followed by an alignable form.